Rework the client registration workflow and client id
TelemetryRepoBranch: client_customer_id_refinements

Previously the client id value in telemetry reports and bundles was
the unique id assigned to a telemery client when it registered with
and upstream telemetry service, whether gateway or relay. With these
changes the client id becomes an UUID string generated by telemetry
clients when they first register with an upstream telemetry server.

Previously when a telemetry client was registering with an upstream
telemetry service it will include a client instance id arbitrary
string value that was intended to uniquely identify the client. With
this update, telemetry clients will instead generated a UUID value,
their client id, to be included in their registration requests, along
with a timestamp and an option systemUUID value. The latter fields
can be used to differentiate between telemetry clients that have
generated the same client id value, in those rare cases where this
happens.

Then result of a regsiter request now includes a unique registration
id value, rather than a client id, and subsequent requests from the
client should include this value as an X-Telemetry-Registration-Id
header value rather than the X-Telemetry-Client-Id header.

The clients table schema has been updated to store the client id,
systemUUID and timestamp values supplied as part of a register
request.

The telemetry table schema has been updated to store the client id
as a string value, rather than an integer.

Updated the GitHub PR testing workflow to allow specifying the
branch of the SUSE/telemetry to use when running tests.